refactoring post salesforce agent asssigned
